requirements.txt
torch
torchvision
pytorch-lightning
tqdm
gdown
wandb
hydra-core
omegaconf
